+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/implementation/models/RecognizeOptionsInternal.java
@@ -56,7 +56,6 @@ public final class RecognizeOptionsInternal {
 
     /**
      * Get the initialSilenceTimeoutInSeconds property: Time to wait for first input after prompt (if any).
-     * If left unspecified, the default value is set to 5 seconds.
      *
      * @return the initialSilenceTimeoutInSeconds value.
      */